docker_project/requirements.txt
Flask==2.2.2
psycopg2==2.9.6
pytest==7.2.0
Werkzeug==2.2.2
pytest-flask==1.2.0
allure-pytest==2.13.5